- Rename to yakworks:eventlog - Rename package to yakworks.eventlog - Remove unused files
