+++ .gitignore
@@ -88,6 +88,7 @@ logs
 # .idea/*.iml
 # .idea/modules
 # CMake
+*.iml
 cmake-build-*/
 # Mongo Explorer plugin
 .idea/**/mongoSettings.xml
