+++ plugin/grails-app/domain/yakworks/eventlog/EventLog.groovy
@@ -2,7 +2,7 @@
 * Copyright 2019 Yak.Works - Licensed under the Apache License, Version 2.0 (the "License")
 * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 */
-package grails.eventlog
+package yakworks.eventlog
 
 import grails.compiler.GrailsCompileStatic
