+++ plugin/src/main/groovy/yakworks/eventlog/GrailsEventlogGrailsPlugin.groovy
@@ -2,11 +2,11 @@
 * Copyright 2019 Yak.Works - Licensed under the Apache License, Version 2.0 (the "License")
 * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 */
-package grails.eventlog
+package yakworks.eventlog
 
 import grails.plugins.Plugin
 
-class GrailsEventLogGrailsPlugin extends Plugin {
+class GrailsEventlogGrailsPlugin extends Plugin {
 
     // the version or versions of Grails the plugin is designed for
     def grailsVersion = "3.3.10 > *"
